scripts/requirements.txt
mysql-connector-python
requests